[Pal/{Linux,Linux-SGX}] Fix signal handling on untrusted stack
First, this commit enables the usage of alternative stack by using
SA_ONSTACK during sigaction() and fixes the bug with ss.ss_sp (it
should point not to the top of the altstack but to its bottom).
Also, the size of the alternative stack is increased to normal
Linux value of 64KB (otherwise stack corruptions ensued).

Second, this commit consolidates the logic of Linux-SGX PAL of
allocating data on the untrusted stack. The normal OCALL handling
used `enclave_tls.ustack` to copy syscall arguments from enclave
memory to untrusted memory and back. On the other hand, the AEX
signal handling used `SGX_GRP.ursp` to operate on the same untrusted
stack. This could clobber the stack if AEX happened during OCALL
processing, and result in corrupted data. This commit consolidates
this logic to always use `SGX_GPR.ursp` and allocate new stack frames
on the untrusted stack.

This commit also adds a corresponding PAL test to validate that
the normal stack and the altstack are different.