[PAL] Use #defines for va_list ap -> va_list VA_AP(ap)
On powerpc64 the function that is called with a 'va_list ap' parameter
must either process the 'ap' parameter itself or must pass it to other
functions as a pointer. So those 2nd level functions need to have a
parameter 'va_list *ap'. This patch now solves this problem using #define's
that we apply to those 2nd level functions signature where the parameter
is now 'va_list VA_AP(ap)'. VA_AP on x86_64 is simply again 'ap' while on
powerpc64 it would be '*ap'. Since powerpc64 needs to be able to create a
pointer to pass to the 2nd level functions, we also introduce VA_AP_PARAM(AP)
that on powerpc64 would be '&(AP)' but on x86_64 can again be 'AP' since no
pointer is needed.