[PAL+LibOS] Fix SIGPIPE emulation
This PR addresses issue #1534.

We must not install a signal handler for SIGPIPE on the PAL layer that
simply forwards all SIGPIPE signals to the LibOS layer since we have
internal file descriptors that also may also receive a SIGPIPE and we
must not forward the SIGPIPE received on those to the application but
handle their EPIPE after the write operation.

The solution is to ignore the SIGPIPE on the PAL layer and intercept
EPIPEs on the LibOS layer after write/writev operations and then do
the signal emulation.